+++ libs/sdk-core/src/lib/db/kb/kb.models.ts
@@ -17,18 +17,6 @@ export enum LabelSetKind {
   PARAGRAPHS = 'PARAGRAPHS',
 }
 
-export enum EventType {
-  VISITED = 'VISITED',
-  MODIFIED = 'MODIFIED',
-  DELETED = 'DELETED',
-  NEW = 'NEW',
-  STARTED = 'STARTED',
-  STOPPED = 'STOPPED',
-  SEARCH = 'SEARCH',
-  PROCESSED = 'PROCESSED',
-  CHAT = 'CHAT',
-}
-
 export interface IKnowledgeBoxCreation {
   id: string;
   slug: string;
@@ -152,20 +140,11 @@ export interface IKnowledgeBox extends IKnowledgeBoxCreation {
 
   suggest(query: string): Observable<Search.Suggestions | IErrorResponse>;
 
-  feedback(answerId: string, good: boolean): Observable<void>;
-  listFeedback(): Observable<string[]>;
-
   counters(): Observable<Counters>;
 
   listResources(page?: number, size?: number): Observable<ResourceList>;
 
   getTempToken(): Observable<string>;
-
-  listActivity(type?: EventType, page?: number, size?: number): Observable<EventList>;
-
-  listActivityDownloads(type: EventType): Observable<ActivityDownloadList>;
-
-  downloadActivity(type: EventType, month: string): Observable<Blob>;
 
   getConfiguration(): Observable<{ [id: string]: any }>;
   getLearningSchema(): Observable<LearningConfigurations>;
@@ -402,10 +381,6 @@ export interface EventList {
 }
 
 export type Event = { [key: string]: any };
-
-export interface ActivityDownloadList {
-  downloads: string[];
-}
 
 export interface ServiceAccountKey {
   id: string;